% Generated by build script
\name{cern_experiments}
\alias{cern_experiments}
\title{List available experiments in CERN Open Data}
\usage{
cern_experiments()
}
\arguments{
\item{size}{Number of records to scan (default 100)}
}
\value{tibble: experiment, count}
\description{List available experiments in CERN Open Data}
